# services/cursor-sim/internal/api/docs/static/openapi/github-sim-api.yaml
openapi: 3.1.0
info:
  title: GitHub Repository Simulation API
  description: |
    Simulated GitHub-style endpoints for PR lifecycle, commit history, and code review data.
    
    These endpoints complement the Cursor AI telemetry API to provide a complete dataset
    for research on AI impact on software development lifecycle (SDLC).
    
    **Research Framework Support:**
    - Independent Variables: PR Volume, PR Scatter, Greenfield Index, Repo Maturity
    - Velocity Metrics: Coding Lead Time, Pickup Time, Review Lead Time, Throughput
    - Review Costs: Review Density, Iteration Count, Rework Ratio, Scope Creep
    - Quality Metrics: Revert Rate, Code Survival, Hotfix Follow-up
    
    **JOIN Keys with Cursor API:**
    - `commit_sha` ↔ `/analytics/ai-code/commits.commitHash`
    - `author_email` ↔ `/analytics/ai-code/commits.userEmail`
    - `repo_name` ↔ `/analytics/ai-code/commits.repoName`
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local simulator
...
paths:
...
  /repos:
    get:
      tags: [Repositories]
      summary: List repositories
      description: Returns all repositories in the simulation with metadata.
      operationId: listRepos
      parameters:
        - $ref: '#/components/parameters/org'
      responses:
        '200':
          description: Repository list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoListResponse'
...
  /research/dataset:
    get:
      tags: [Research]
      summary: Export research dataset
      description: |
        Returns a pre-joined dataset with all research variables per PR/commit.
        
        This endpoint provides a single table optimized for statistical analysis,
        joining Cursor AI telemetry with GitHub PR lifecycle data.
        
        **Columns include:**
        - Independent: ai_lines_added, ai_lines_deleted, pr_volume, pr_scatter, greenfield_index
        - Velocity: coding_lead_time_hours, pickup_time_hours, review_lead_time_hours
        - Review Costs: review_density, iteration_count, rework_ratio, scope_creep
        - Quality: is_reverted, survival_rate_30d, has_hotfix_followup
      operationId: getResearchDataset
      parameters:
        - $ref: '#/components/parameters/since'
        - $ref: '#/components/parameters/until'
        - name: granularity
          in: query
          description: Level of aggregation
          schema:
            type: string
            enum: [commit, pr, developer_day, developer_week]
            default: pr
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, parquet]
            default: json
      responses:
        '200':
          description: Research dataset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchDataset'
            text/csv:
              schema:
                type: string
                format: binary
...
components:
  parameters:
...
    org:
      name: org
      in: query
      description: Filter by organization
      schema:
        type: string
    since:
      name: since
      in: query
      description: Start date (ISO 8601)
      schema:
        type: string
        format: date-time
    until:
      name: until
      in: query
      description: End date (ISO 8601)
      schema:
        type: string
        format: date-time
...
  schemas:
...
    RepoListResponse:
      type: object
      properties:
        repositories:
          type: array
          items:
            $ref: '#/components/schemas/RepositorySummary'
        total_count:
          type: integer
    RepositorySummary:
      type: object
      properties:
        full_name:
          type: string
          example: "acme-corp/payment-service"
        default_branch:
          type: string
          example: "main"
        primary_language:
          type: string
          example: "TypeScript"
...
    ResearchDataset:
      type: object
      description: |
        Pre-joined dataset for statistical analysis.
        One row per PR (or per specified granularity).
      properties:
        granularity:
          type: string
          enum: [commit, pr, developer_day, developer_week]
        columns:
          type: array
          items:
            type: string
          description: Column names in order
          example: [
            "pr_number", "author_email", "repo_name",
            "ai_lines_added", "ai_lines_deleted", "non_ai_lines_added",
            "pr_volume", "pr_scatter", "greenfield_index",
            "coding_lead_time_hours", "pickup_time_hours", "review_lead_time_hours",
            "review_density", "iteration_count", "rework_ratio", "scope_creep",
            "is_reverted", "survival_rate_30d", "has_hotfix_followup"
          ]
        rows:
          type: array
          items:
            type: array
            items: {}
        total_rows:
          type: integer
...